Refacto playbook manager
